+++ lib/client.js
@@ -1,6 +1,7 @@
 "use strict";
 
-var _ = require("./utils");
+var decorator = require("./decorator"),
+    _ = require("./utils");
 
 var protocols = {
   http: require("./drivers/http")
@@ -18,46 +19,6 @@ function checkArgs(protocol, opts) {
   if (opts.port == null) { throw new Error("No port option provided"); }
 }
 
-function process(driver, data) {
-  function bindCommand(target, path) {
-    return function(name) {
-      if (target[name]) { return; }
-      target[name] = driver.command.bind(driver, name, path);
-    };
-  }
-
-  function on(path) {
-    return function(event, callback) {
-      return driver.event(event, path, callback);
-    };
-  }
-
-  data.commands.forEach(bindCommand(data, ""));
-  data.on = on("");
-  data.robot = function(name) {
-    return data.robots.filter(function(bot) { return bot.name === name; })[0];
-  };
-
-  data.robots.forEach(function(robot) {
-    var path = "robots/" + robot.name;
-
-    robot.commands.forEach(bindCommand(robot, path));
-    robot.on = on(path);
-    robot.device = function(name) {
-      return robot.devices.filter(function(d) { return d.name === name; })[0];
-    };
-
-    robot.devices.forEach(function(device) {
-      var devicePath = path + "/devices/" + device.name;
-
-      device.commands.forEach(bindCommand(device, devicePath));
-      device.on = on(devicePath);
-    });
-  });
-
-  return data;
-}
-
 var Client = module.exports = function Client(protocol, opts) {
   checkArgs(protocol, opts);
 
@@ -73,7 +34,21 @@ Client.prototype.connect = function(callback) {
   this.driver.connect(function(err, data) {
     if (err) { return _.resolve(callback, err); }
     data = _.extend({}, data);
-    _.resolve(callback, null, process(driver, data));
+    _.resolve(callback, null, decorator.decorate(driver, data));
+  });
+};
+
+Client.prototype.update = function(callback) {
+  var driver = this.driver;
+
+  if (!driver.connected) {
+    return _.resolve(callback, "Not connected. Call Client#connect instead.");
+  }
+
+  driver.update(function(err, data) {
+    if (err) { return _.resolve(callback, err); }
+    data = _.extend({}, data);
+    _.resolve(callback, null, decorator.decorate(driver, data));
   });
 };
